nixos-secrets: rename from `nixos-vars`
