+++ pkgs/by-name/ni/nixos-secrets/example/config/key-target.txt
