+++ pkgs/by-name/ni/nixos-secrets/src/nixos_secrets/__main__.py
