+++ pkgs/by-name/ni/nixos-secrets/src/nixos_secrets/error.py
@@ -1,0 +1,4 @@
+class SecretsError(Exception):
+    """Base exception for nixos-secrets errors."""
+
+    pass
