--- pkgs/by-name/ni/nixos-vars/example/flake/flake.nix
@@ -1,21 +1,0 @@
-{
-  inputs.nixpkgs.url = "github:starlitcanopy/nixpkgs?ref=master";
-
-  outputs = inputs: {
-    nixosConfigurations.example = inputs.nixpkgs.lib.nixosSystem {
-      system = "aarch64-linux";
-
-      modules = [ ../config ];
-    };
-
-    # Example of how one can override the host package set!
-    varsConfigurations.differentArch =
-      let
-        pkgsHost = inputs.nixpkgs.legacyPackages.x86_64-linux;
-      in
-      pkgsHost.nixos-vars.jsonify {
-        inherit pkgsHost;
-        configuration = inputs.self.nixosConfigurations.example;
-      };
-  };
-}
