--- pkgs/by-name/ni/nixos-vars/package.nix
@@ -1,55 +1,0 @@
-{
-  lib,
-  mkShell,
-  nixosTests,
-  python3,
-  python3Packages,
-  ruff,
-  makeWrapper,
-  bubblewrap,
-}:
-
-python3Packages.buildPythonApplication {
-  __structuredAttrs = true;
-
-  name = "nixos-vars";
-  format = "pyproject";
-  nativeBuildInputs = [
-    python3Packages.setuptools
-    makeWrapper
-  ];
-  src = ./src;
-
-  postFixup = ''
-    wrapProgram $out/bin/nixos-vars \
-      --prefix PATH : ${bubblewrap}/bin
-  '';
-
-  passthru.devShell = mkShell {
-    packages = [
-      python3
-      bubblewrap
-      ruff
-    ];
-  };
-
-  passthru.tests = {
-    inherit (nixosTests)
-      nixos-vars-basic-generators
-      ;
-  };
-
-  # This is perhaps not a good idea?
-  passthru.jsonify = import ./src/nixos_vars/nix/jsonify.nix;
-
-  meta = {
-    description = "NixOS secret management abstraction";
-    homepage = "https://github.com/NixOS/nixpkgs/tree/master/pkgs/by-name/ni/nixos-vars";
-    license = lib.licenses.mit;
-    maintainers = with lib.maintainers; [
-      lassulus
-      prescientmoon
-    ];
-    mainProgram = "nixos-vars";
-  };
-}
