--- pkgs/by-name/ni/nixos-vars/src/nixos_vars/cli.py
@@ -1,131 +1,0 @@
-import argparse
-import sys
-from pathlib import Path
-from .error import VarsError
-from .eval import evaluate_config
-from .generate import generate_vars
-from .gc import collect_garbage
-from .args import VarsArgs
-from .deploy import deploy
-
-
-def common_args(parser: argparse.ArgumentParser):
-    parser.add_argument(
-        "-f",
-        "--file",
-        type=Path,
-        default=None,
-        metavar="<file>",
-        help="Path to config.nix",
-    )
-
-    parser.add_argument(
-        "-F",
-        "--flake",
-        type=str,
-        default=None,
-        metavar="<flake-uri#name>",
-        help="Flake-based alternative to --file",
-    )
-
-    parser.add_argument(
-        "-j",
-        "--json",
-        type=str,
-        default=None,
-        metavar="<file>",
-        help="Parse the vars configuration from the given JSON file",
-    )
-
-    parser.add_argument(
-        "-A",
-        "--attr",
-        type=str,
-        metavar="<attr-path>",
-        help="Gets the attribute at the given path",
-    )
-
-    parser.add_argument(
-        "--dry-run",
-        action="store_true",
-        help="Print what would be done without executing",
-    )
-
-    parser.add_argument(
-        "--no-sandbox",
-        action="store_true",
-        dest="disable_sandbox",
-        help="Do not run the generator scripts inside bubblewrap",
-    )
-
-    parser.add_argument(
-        "-v",
-        "--verbose",
-        action="store_true",
-        help="Print all the logs generated by the various scripts",
-    )
-
-
-def main() -> None:
-    parser = argparse.ArgumentParser(description="vars-ng-ng CLI")
-
-    # Global arguments
-    subparsers = parser.add_subparsers(title="commands", dest="command", required=True)
-
-    eval_parser = subparsers.add_parser("evaluate")
-
-    common_args(eval_parser)
-    gen_parser = subparsers.add_parser("generate", help="(Re)generate secrets")
-    common_args(gen_parser)
-    gen_parser.add_argument(
-        "-g",
-        "--generator",
-        dest="generators",
-        metavar="<generator>",
-        type=str,
-        default=[],
-        action="append",
-        help="Generator(s) to force the regeneration of.",
-    )
-    gen_parser.add_argument(
-        "-s",
-        "--set",
-        dest="set",
-        metavar="<generator=path>",
-        type=str,
-        default=[],
-        action="append",
-        help="Generator(s) to force the regeneration of.",
-    )
-
-    gc_parser = subparsers.add_parser("collect-garbage", help="Delete stale secrets")
-
-    common_args(gc_parser)
-    deploy_parser = subparsers.add_parser("deploy", help="Deploy secrets")
-    common_args(deploy_parser)
-    deploy_parser.add_argument(
-        "-l",
-        "--local",
-        type=str,
-        metavar="<sytem-root>",
-        help="Deploy to a locally attached filesystem",
-    )
-
-    args = parser.parse_args()
-    args = VarsArgs.from_dict(vars(args))
-
-    try:
-        config = evaluate_config(args)
-        if args.command == "evaluate":
-            print(config)
-        elif args.command == "generate":
-            generate_vars(args, config)
-        elif args.command == "collect-garbage":
-            collect_garbage(args, config)
-        elif args.command == "deploy":
-            deploy(args, config)
-        else:
-            raise VarsError(f"Command '{args.command}' is not implemented :(")
-    except VarsError as e:
-        print(str(e), file=sys.stderr)
-        exit(1)
