--- pkgs/by-name/ni/nixos-vars/src/nixos_vars/error.py
@@ -1,4 +1,0 @@
-class VarsError(Exception):
-    """Base exception for vars-ng errors."""
-
-    pass
